+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - 2025-05-20
 
 ### Added
+* Added inputs for disabling failure on decrease when coverage is above a specified limit:
+  * `coverage-decrease-limit`: Overrides `fail-on-coverage-decrease` when coverage is at least this limit
+  * `branches-decrease-limit`: Overrides `fail-on-branches-decrease` when branches coverage is at least this limit
 
 ### Changed